Update extension.js
Removed import of import.ui.tweener. The imported constant didn't appear to be referenced anywhere in the code and Gnome 3.38 changed the location to import.tweener.tweener. Because there don't appear to be any references, I have removed the import instead of getting into the complexity of checking Gnome version to determine which import to use.

Added a constant gnome_at_least_3_38 which is used later in the createOverlay function.

Updated the createOverlay function so that it checks gnome_at_least_3_38 to determine which syntax to use to call new StBin() to create the new overlay.

I've tested Gnome 3.38 on ArchLinux  EndeavourOS and can't find anything that isn't working as expected.

PS - You might want to update metadata.json to add Gnome 3.38 to the version list.